perf(layout): seperate layout
